% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{addAreaSelect}
\alias{addAreaSelect}
\alias{addAreaSelectControl}
\alias{removeAreaSelect}
\alias{setAreaSelectDimensions}
\title{Adds a resizable centered box on top of the map letting users select a square area (bounding box).}
\usage{
addAreaSelect(map, width = 200, height = 300, keepAspectRatio = FALSE)

addAreaSelectControl(map, position = c("topleft", "topright", "bottomleft",
  "bottomright"), width = 200, height = 300, keepAspectRatio = FALSE)

setAreaSelectDimensions(map, width, height)

removeAreaSelect(map)
}
\arguments{
\item{map}{The map widget.}

\item{width}{Width in pixels.}

\item{keepAspectRatio}{if set to TRUE the aspect ratio of the selection box can be changed.}

\item{position}{The position of the control button.}

\item{heigth}{Height in pixels.}
}
\description{
Adds a resizable centered box on top of the map letting users select a square area (bounding box).

Adds a button to dynamically add/remove area selection box.

Changes the dimension of the area selection box.

Removes the area selection box.
}
\section{Functions}{
\itemize{
\item \code{addAreaSelectControl}: shows a button to add/remove area selection box

\item \code{setAreaSelectDimensions}: change dimensions of the selection box.

\item \code{removeAreaSelect}: removes the area selection box.
}}

