% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{addBootstrap}
\alias{addBootstrap}
\title{Required for using BootStrap Fonts}
\usage{
addBootstrap(map)
}
\arguments{
\item{map}{the map to add awesome Markers to.}
}
\description{
Required for using BootStrap Fonts
}

