% Please edit documentation in R/plugin-gps.R
\name{addControlGPS}
\alias{addControlGPS}
\title{Add a gps to the Map
\url{https://github.com/Norkart/Leaflet-MiniMap}}
\usage{
addControlGPS(map)
}
\arguments{
\item{map}{a map widget object}
}
\description{
Add a gps to the Map
\url{https://github.com/Norkart/Leaflet-MiniMap}
}
