% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{addCustomLegend}
\alias{addCustomLegend}
\title{Add a custom html legend to the map}
\usage{
addCustomLegend(map, html, position = c("topright", "bottomright",
  "bottomleft", "topleft"), className = "info legend", layerId = NULL,
  group = NULL)
}
\arguments{
\item{map}{a map widget object created from \code{\link{leaflet}()}}

\item{html}{the html code to render the legend}

\item{position}{the position of the legend}

\item{className}{extra CSS classes to append to the control, space separated}

\item{layerId}{the ID of the legend; subsequent calls to \code{addLegend}
or \code{addControl} with the same \code{layerId} will replace this
legend. The ID can also be used with \code{removeControl}.}

\item{group}{\code{group} name of a leaflet layer group.
Supplying this value will tie the legend to the leaflet layer group
with this name and will auto add/remove the legend as the
group is added/removed, for example via layerControl.
You will need to set the \code{group} when you add a layer
(e.g. \code{\link{addPolygons}}) and supply the same name here.}
}
\description{
Add a custom html legend to the map
}
