% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-easybutton.R
\name{addEasyButton}
\alias{addEasyButton}
\title{Add a EasyButton on the map
see \url{https://github.com/CliffCloud/Leaflet.EasyButton}}
\usage{
addEasyButton(map, button = easyButton())
}
\arguments{
\item{map}{a map widget object}

\item{button}{the button object created with \code{\link{easyButton}}}
}
\description{
Add a EasyButton on the map
see \url{https://github.com/CliffCloud/Leaflet.EasyButton}
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addEasyButton(easyButton(
     icon = htmltools::span(class='star','&starf;'),
     onClick = JS("function(btn, map){ map.setZoom(1);}")))

}

