% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-easybutton.R
\name{addEasyButtonBar}
\alias{addEasyButtonBar}
\title{Add a easyButton bar on the map
see \url{https://github.com/CliffCloud/Leaflet.EasyButton}}
\usage{
addEasyButtonBar(map, buttons, position = "topleft", id = NULL)
}
\arguments{
\item{map}{a map widget object}

\item{buttons}{the buttons object created with \code{\link{easyButtonList}}}

\item{position}{topleft|topright|bottomleft|bottomright}

\item{id}{id for the button bar}
}
\description{
Add a easyButton bar on the map
see \url{https://github.com/CliffCloud/Leaflet.EasyButton}
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addEasyButtonBar(easyButtonList(
   easyButton(
     icon = htmltools::span(class='star','&starf;'),
     onClick = JS("function(btn, map){ alert('Button 1');}")),
   easyButton(
     icon = htmltools::span(class='star','&target;'),
     onClick = JS("function(btn, map){ alert('Button 2');}"))))


}

