% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/events.R
\name{addEventListener}
\alias{addEventListener}
\title{Methods to add event listener map widget}
\usage{
addEventListener(map, ...)
}
\arguments{
\item{map}{a map widget object created from \code{\link{leaflet}()}}

\item{...}{any arguments with names equal to event name, and value to javascript function. See example.}
}
\value{
The modified map widget.
}
\description{
A series of methods to manipulate the map.
}
\examples{
library(leaflet)
m = leaflet() \%>\% addTiles() \%>\% setView(-93.65, 42.0285, zoom = 17) \%>\%
 addEventListener(click ="function(e) {alert('click');}",
 zoomend ="function(e) {alert('zoomend');}")
 m
}
\references{
\url{http://leafletjs.com/reference.html#map-click}
}

