% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{addFontAwesome}
\alias{addFontAwesome}
\title{Required for using Font-Awesome Fonts}
\usage{
addFontAwesome(map)
}
\arguments{
\item{map}{the map to add awesome Markers to.}
}
\description{
Required for using Font-Awesome Fonts
}

