% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-graticule.R
\name{addGraticule}
\alias{addGraticule}
\title{Add a Graticule on the map
see \url{https://github.com/turban/Leaflet.Graticule}}
\usage{
addGraticule(map, interval = 20, style = list(color = "#333", weight = 1),
  layerId = NULL, group = NULL)
}
\arguments{
\item{map}{a map widget object}

\item{interval}{The spacing in map units between horizontal and vertical lines.}

\item{style}{path options for the generated lines. See \url{http://leafletjs.com/reference.html#path-options}}

\item{layerId}{the layer id}

\item{group}{the name of the group this layer belongs to.}
}
\description{
Add a Graticule on the map
see \url{https://github.com/turban/Leaflet.Graticule}
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addGraticule()

}

