% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{addIonIcon}
\alias{addIonIcon}
\title{Required for using IonIcon Fonts}
\usage{
addIonIcon(map)
}
\arguments{
\item{map}{the map to add awesome Markers to.}
}
\description{
Required for using IonIcon Fonts
}

