% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{addLocationFilter}
\alias{addLocationFilter}
\alias{removeLocationFilter}
\title{Adds a  draggable/resizable rectangle on top of the map for location filtering.}
\usage{
addLocationFilter(map, position = c("topleft", "topright", "bottomleft",
  "bottomright"), enable = FALSE, sw_lat = NULL, sw_lng = NULL,
  ne_lat = NULL, ne_lng = NULL)

removeLocationFilter(map)
}
\arguments{
\item{map}{The map widget.}

\item{position}{The position of the control button.}

\item{enable}{If set to TRUE the location filter is enabled when added to the map.}

\item{sw_lat}{Latitude of South-West End of the filter box.}

\item{sw_lng}{Longitude of South-West End of the filter box.}

\item{ne_lat}{Latitude of North-East End of the filter box.}

\item{ne_lng}{Longitude of North-East End of the filter box.}
}
\description{
Adds a  draggable/resizable rectangle on top of the map for location filtering.

Removes the location filter box.
}
\section{Functions}{
\itemize{
\item \code{removeLocationFilter}: removes the location filter box.
}}

