% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-minimap.R
\name{addMiniMap}
\alias{addMiniMap}
\title{Add a minimap to the Map}
\usage{
addMiniMap(map, position = "bottomright", width = 150, height = 150,
  collapsedWidth = 19, collapsedHeight = 19, zoomLevelOffset = -5,
  zoomLevelFixed = NULL, zoomAnimation = FALSE, toggleDisplay = FALSE,
  autoToggleDisplay = NULL)
}
\arguments{
\item{map}{a map widget object}

\item{position,width,height,collapsedWidth,collapsedHeight;}{see
\url{https://github.com/Norkart/Leaflet-MiniMap}}

\item{zoomLevelOffset,zoomLevelFixed,zoomAnimation,toggleDisplay,autoToggleDisplay;}{see
\url{https://github.com/Norkart/Leaflet-MiniMap}}
}
\description{
Add a minimap to the Map
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addMiniMap()
}

