% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-Proj4Leaflet.R
\name{addProj4Leaflet}
\alias{addProj4Leaflet}
\title{Adds support for Proj4Leaflet}
\usage{
addProj4Leaflet(map)
}
\arguments{
\item{map}{The Leaflet map.}
}
\description{
Adds support for Proj4Leaflet
}

