% Please edit documentation in R/plugin-timeDimension.R
\name{addRasterTime}
\alias{addRasterTime}
\title{Add Raster with Time Dimension}
\usage{
addRasterTime(map, x, colors = "Spectral", opacity = 1, attribution = NULL, 
    layerId = NULL, group = NULL, project = TRUE, maxBytes = 4 * 1024 * 1024, 
    tz = NULL)
}
\description{
Add Raster with Time Dimension
}
