% Please edit documentation in R/plugin-search.R
\name{addSearch}
\alias{addSearch}
\title{Add a search control to the map.}
\usage{
addSearch(map, layer = "search", position = "topright", initial = FALSE)
}
\arguments{
\item{map}{a map widget object}

\item{layer}{a searchable layer}

\item{position}{standard \href{http://leafletjs.com/reference.html#control-positions}{Leaflet control position options}.}

\item{initial}{initial...}
}
\value{
modified map
}
\description{
Add a search control to the map.
}
