% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-simplegraticule.R
\name{addSimpleGraticule}
\alias{addSimpleGraticule}
\title{Add a simple Graticule on the map}
\usage{
addSimpleGraticule(map, interval = 20, showOriginLabel = TRUE,
  redraw = "move", layerId = NULL, group = NULL)
}
\arguments{
\item{map}{a map widget object}

\item{interval,showOriginLabel,redraw}{see \url{https://github.com/ablakey/Leaflet.SimpleGraticule}}

\item{layerId}{the layer id}

\item{group}{the name of the group this layer belongs to.}
}
\description{
Add a simple Graticule on the map
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addSimpleGraticule()
}

