% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breweries91}
\alias{breweries91}
\title{Selected breweries in Franconia}
\format{\code{sp::SpatialPointsDataFrame}}
\description{
Selected breweries in Franconia (zip code starting with 91...)
}
\details{
This dataset contains selected breweries in Franconia. It is a
subset of a larger database that was compiled by students at the
University of Marburg for a seminar called
"The Geography of Beer, sustainability in the food industry"
}

