% Please edit documentation in R/layers.R
\name{setStyleGeoJSON}
\alias{setStyleGeoJSON}
\title{Change style of GeoJSON feature}
\usage{
setStyleGeoJSON(map, layerId = NULL, attribute = "id", value = NULL, command = "=", 
    style = NULL)
}
\arguments{
\item{style}{in JSON}
}
\section{Functions}{
\itemize{
\item \code{setStyleGeoJSON}: style of GeoJSON feature
}}
