% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createLeafletMap}
\alias{createLeafletMap}
\title{Create a Leaflet map object in R}
\usage{
createLeafletMap(session, outputId)
}
\arguments{
\item{session}{The \code{session} argument passed through from the
\code{\link[shiny]{shinyServer}} server function.}

\item{outputId}{The string identifier that was passed to the corresponding
\code{\link{leafletMap}()}.}
}
\value{
A list of methods. See the package vignette \code{vignette('intro',
  'leaflet'} for details.
}
\description{
This function is called from \file{server.R} and returns an object that can
be used to manipulate the Leaflet map from R.
}

