% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{derivePoints}
\alias{derivePoints}
\alias{derivePolygons}
\title{Normalize data to a data frame or Spatial object.}
\usage{
derivePoints(
  data,
  lng = NULL,
  lat = NULL,
  missingLng = missing(lng),
  missingLat = missing(lat),
  funcName = "f"
)

derivePolygons(
  data,
  lng = NULL,
  lat = NULL,
  missingLng = missing(lng),
  missingLat = missing(lat),
  funcName = "f"
)
}
\arguments{
\item{data}{map data}

\item{lng, lat}{longitude and latitude}

\item{missingLng}{whether lng is missing}

\item{missingLat}{whether lat is missing}

\item{funcName}{Name of calling function (for logging)}
}
\value{
* `derivePoints()` returns a lng/lat data frame.
* `derivePolygons()` returns a list with bbox attributes.
}
\description{
Given a data object and lng/lat arguments (which may be `NULL`)
[meaning infer from data], formula [which should be evaluated with respect to the data], or
vector data [which should be used as-is]) return a lng/lat data frame.
}
\keyword{internal}
