% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{derivePolygons}
\alias{derivePolygons}
\title{Given a data object and lng/lat arguments (which may be NULL [meaning infer
from data], formula [which should be evaluated with respect to the data], or
vector data [which should be used as-is]) return a spatial object}
\usage{
derivePolygons(
  data,
  lng = NULL,
  lat = NULL,
  missingLng = missing(lng),
  missingLat = missing(lat),
  funcName = "f"
)
}
\arguments{
\item{data}{map data}

\item{lng}{longitude}

\item{lat}{latitude}

\item{missingLng}{whether lng is missing}

\item{missingLat}{whether lat is missing}

\item{funcName}{Name of calling function (for logging)}
}
\description{
Given a data object and lng/lat arguments (which may be NULL [meaning infer
from data], formula [which should be evaluated with respect to the data], or
vector data [which should be used as-is]) return a spatial object
}
