% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-draw.R
\name{drawPolylineOptions}
\alias{drawCircleOptions}
\alias{drawMarkerOptions}
\alias{drawPolygonOptions}
\alias{drawPolylineOptions}
\alias{drawRectangleOptions}
\title{Customize draw plugin behavior}
\usage{
drawPolylineOptions(allowIntersection = TRUE, drawError = list(color =
  "#b00b00", timeout = 2500), guidelineDistance = 20,
  shapeOptions = list(stroke = TRUE, color = "#f06eaa", weight = 4, opacity =
  0.5, fill = FALSE, clickable = TRUE), metric = TRUE, zIndexOffset = 2000,
  repeatMode = FALSE)

drawPolygonOptions(allowIntersection = TRUE, drawError = list(color =
  "#b00b00", timeout = 2500), guidelineDistance = 20,
  shapeOptions = list(stroke = TRUE, color = "#f06eaa", weight = 4, opacity =
  0.5, fill = TRUE, fillColor = NULL, fillOpacity = 0.2, clickable = TRUE),
  metric = TRUE, zIndexOffset = 2000, repeatMode = FALSE,
  showArea = FALSE)

drawRectangleOptions(shapeOptions = list(stroke = TRUE, color = "#f06eaa",
  weight = 4, opacity = 0.5, fill = TRUE, fillColor = NULL, fillOpacity = 0.2,
  clickable = TRUE), metric = TRUE, repeatMode = FALSE)

drawCircleOptions(shapeOptions = list(stroke = TRUE, color = "#f06eaa", weight
  = 4, opacity = 0.5, fill = TRUE, fillColor = NULL, fillOpacity = 0.2,
  clickable = TRUE), showRadius = TRUE, metric = TRUE, repeatMode = FALSE)

drawMarkerOptions(icon = NULL, zIndexOffset = 2000, repeatMode = FALSE)
}
\arguments{
\item{allowIntersection}{determines if line segments can cross}

\item{drawError}{configuration options for the error that displays if an
intersection is detected}

\item{guidelineDistance}{distance in pixels between each guide dash}

\item{shapeOptions}{the options used when drawing the polyline/polygon on the
map}

\item{metric}{determines which measurement system (metric or imperial) is
used}

\item{zIndexOffset}{this should be a high number to ensure that you can draw
over all other layers on the map}

\item{repeatMode}{determines if the draw tool remains enabled after drawing a
  shape}

\item{showArea}{Show the area of the drawn polygon. \strong{The area is only
approximate and become less accurate the larger the polygon is.}}

\item{showRadius}{whether to show the radius of the drawn circle}

\item{icon}{a custom icon, as created by \code{\link{makeIcon}}}
}
\description{
Constructors for options objects that can be passed as the corresponding
parameter to \code{\link{addDrawToolbar}}.
}

