% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-easybutton.R
\name{easyButton}
\alias{easyButton}
\title{Creates an easy button.
see \url{https://github.com/CliffCloud/Leaflet.EasyButton}}
\usage{
easyButton(icon = htmltools::span(class = "easy-button", "!"),
  title = "Easy Button",
  onClick = JS("function(btn, map){alert('That was easy!');}"),
  position = "topleft", id = NULL, states = NULL)
}
\arguments{
\item{icon}{the button icon}

\item{title}{text to show on hover}

\item{onClick}{the action to take}

\item{position}{topleft|topright|bottomleft|bottomright}

\item{id}{id for the button}

\item{states}{the states}
}
\description{
Creates an easy button.
see \url{https://github.com/CliffCloud/Leaflet.EasyButton}
}

