% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-easybutton.R
\name{easyButtonList}
\alias{easyButtonList}
\title{Creates a list of easy buttons.}
\usage{
easyButtonList(...)
}
\arguments{
\item{...}{icons created from \code{\link{easyButton}()}}
}
\description{
Creates a list of easy buttons.
}

