% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-easybutton.R
\name{easyButtonState}
\alias{easyButtonState}
\title{Create an easyButton statestate}
\usage{
easyButtonState(stateName, icon, title, onClick)
}
\arguments{
\item{stateName}{the name of the state}

\item{icon}{the button icon}

\item{title}{text to show on hover}

\item{onClick}{the action to take}
}
\description{
Create an easyButton statestate
}

