% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-easybutton.R
\name{easyButtonStateList}
\alias{easyButtonStateList}
\title{Create a list of easyButton states.}
\usage{
easyButtonStateList(...)
}
\arguments{
\item{...}{states created from \code{\link{easyButtonState}()}}
}
\description{
Create a list of easyButton states.
}

