% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{evalFormula}
\alias{evalFormula}
\title{Evaluate list members that are formulae, using the map data as the environment}
\usage{
evalFormula(list, data)
}
\arguments{
\item{list}{with members as formulae}

\item{data}{map data}
}
\description{
Evaluates if provided. Otherwise, the formula environment is evaluated.
}
\keyword{internal}
