% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{expandLimits}
\alias{expandLimits}
\title{Notifies the map of new latitude/longitude of items of interest on the map}
\usage{
expandLimits(map, lat, lng)
}
\arguments{
\item{map}{map object}

\item{lat}{vector of latitudes}

\item{lng}{vector of longitudes}
}
\description{
Notifies the map of new latitude/longitude of items of interest on the map
}
