% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{expandLimitsBbox}
\alias{expandLimitsBbox}
\title{Notifies the map of polygons of interest on the map}
\usage{
expandLimitsBbox(map, poly)
}
\arguments{
\item{map}{map object}

\item{poly}{A spatial object representing a polygon.}
}
\description{
Same as \code{expandLimits()}, but takes a polygon (that presumably has a bbox attr)
rather than lat/lng.
}
