% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterNULL}
\alias{filterNULL}
\title{Remove NULL elements from a list}
\usage{
filterNULL(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A list with `NULL` elements filtered out.
}
\description{
Remove NULL elements from a list
}
\keyword{internal}
