% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadmCHE}
\alias{gadmCHE}
\title{Administrative borders of Switzerland (level 1)}
\format{
\code{sp::SpatialPolygonsDataFrame}
}
\source{
\url{https://gadm.org}
}
\description{
Administrative borders of Switzerland (level 1)
}
\details{
This dataset comes from \url{https://gadm.org}.
It was downloaded using \code{\link[=getData]{getData()}}.
}
\seealso{
Other built in datasets: 
\code{\link{atlStorms2005}},
\code{\link{breweries91}}
}
\concept{built in datasets}
