% Generated by roxygen2 (4.0.2): do not edit by hand
\name{leaflet-exports}
\alias{\%>\%}
\alias{leaflet-exports}
\title{Re-exported objects from other packages}
\description{
This page is only for the purpose of \command{R CMD check}, because all
exported objects must be documented, even if they are only objects
re-exported from other packages, where they have been documented.
}

