% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet-package.R
\docType{package}
\name{leaflet-package}
\alias{leaflet-package}
\title{leaflet: Create Interactive Web Maps with the JavaScript 'Leaflet' Library}
\description{
Create and customize interactive maps using the 'Leaflet' JavaScript library and the 'htmlwidgets' package. These maps can be used directly from the R console, from 'RStudio', in Shiny applications and R Markdown documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rstudio.github.io/leaflet/}
  \item \url{https://github.com/rstudio/leaflet}
  \item Report bugs at \url{https://github.com/rstudio/leaflet/issues}
}

}
\author{
\strong{Maintainer}: Joe Cheng \email{joe@posit.co}

Authors:
\itemize{
  \item Barret Schloerke \email{barret@posit.co} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Bhaskar Karambelkar
  \item Yihui Xie
}

Other contributors:
\itemize{
  \item Hadley Wickham [contributor]
  \item Kenton Russell [contributor]
  \item Kent Johnson [contributor]
  \item Vladimir Agafonkin (Leaflet library) [contributor, copyright holder]
  \item CloudMade (Leaflet library) [copyright holder]
  \item Leaflet contributors (Leaflet library) [contributor]
  \item Brandon Copeland (leaflet-measure plugin) [contributor, copyright holder]
  \item Joerg Dietrich (Leaflet.Terminator plugin) [contributor, copyright holder]
  \item Benjamin Becquet (Leaflet.MagnifyingGlass plugin) [contributor, copyright holder]
  \item Norkart AS (Leaflet.MiniMap plugin) [contributor, copyright holder]
  \item L. Voogdt (Leaflet.awesome-markers plugin) [contributor, copyright holder]
  \item Daniel Montague (Leaflet.EasyButton plugin) [contributor, copyright holder]
  \item Kartena AB (Proj4Leaflet plugin) [contributor, copyright holder]
  \item Robert Kajic (leaflet-locationfilter plugin) [contributor, copyright holder]
  \item Mapbox (leaflet-omnivore plugin) [contributor, copyright holder]
  \item Michael Bostock (topojson) [contributor, copyright holder]
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
