% Generated by roxygen2 (4.0.2): do not edit by hand
\name{leaflet}
\alias{leaflet}
\title{Create a Leaflet map widget}
\usage{
leaflet(id = NULL, width = NULL, height = NULL, padding = 0)
}
\arguments{
\item{id}{a character string as the identifier of the map (you do not need to
provide it unless you want to manipulate the map later in Shiny)}

\item{width}{the width of the map}

\item{height}{the height of the map}

\item{padding}{the padding of the map}
}
\value{
A HTML widget object.
}
\description{
This function creates a Leaflet map widget using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}

