% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leafletCRS}
\alias{leafletCRS}
\title{creates a custom CRS
Refer to \url{https://kartena.github.io/Proj4Leaflet/api/} for details.}
\usage{
leafletCRS(crsClass = "L.CRS.EPSG3857", code = NULL, proj4def = NULL,
  projectedBounds = NULL, origin = NULL, transformation = NULL,
  scales = NULL, resolutions = NULL, bounds = NULL, tileSize = NULL)
}
\arguments{
\item{crsClass}{One of L.CRS.EPSG3857, L.CRS.EPSG4326, L.CRS.EPSG3395,
L.CRS.Simple, L.Proj.CRS, L.Proj.CRS.TMS}

\item{code}{CRS identifier}

\item{proj4def}{Proj4 string}

\item{projectedBounds}{Only when crsClass = 'L.Proj.CRS.TMS'}

\item{origin}{Origin in projected coordinates, if set overrides transformation option.}

\item{transformation}{to use when transforming projected coordinates into pixel coordinates}

\item{scales}{Scale factors (pixels per projection unit, for example pixels/meter)
for zoom levels; specify either scales or resolutions, not both}

\item{resolutions}{factors (projection units per pixel, for example meters/pixel)
for zoom levels; specify either scales or resolutions, not both}

\item{bounds}{Bounds of the CRS, in projected coordinates; if defined,
Proj4Leaflet will use this in the getSize method, otherwise
defaulting to Leaflet's default CRS size}

\item{tileSize}{Tile size, in pixels, to use in this CRS (Default 256)
Only needed when crsClass = 'L.Proj.CRS.TMS'}
}
\description{
creates a custom CRS
Refer to \url{https://kartena.github.io/Proj4Leaflet/api/} for details.
}

