% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\docType{data}
\name{leafletDependencies}
\alias{leafletDependencies}
\title{Various leaflet dependency functions for use in downstream packages}
\format{An object of class \code{list} of length 13.}
\usage{
leafletDependencies
}
\description{
Various leaflet dependency functions for use in downstream packages
}
\examples{
\dontrun{
addBootStrap <- function(map) {
  map$dependencies <- c(map$dependencies, leafletDependencies$bootstrap())
  map
}
}
}
\keyword{datasets}

