% Generated by roxygen2 (4.0.2): do not edit by hand
\name{leafletMap}
\alias{leafletMap}
\title{Create a \code{div} element for a Leaflet map}
\usage{
leafletMap(outputId, width, height,
  initialTileLayer = "http://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
  initialTileLayerAttribution, options = NULL)
}
\arguments{
\item{outputId}{the id of the \samp{<div>} element}

\item{width,height}{The width and height of the map. They can either take a
CSS length (e.g. \code{400px} or \code{50\%}) or a numeric value which will
be interpreted as pixels.}

\item{initialTileLayer}{The URL template for the initial layer of tile images
(the OpenStreetMap tiles are used by default). See
\url{http://leafletjs.com/reference.html#tilelayer} for information about
providing tile layer URLs.}

\item{initialTileLayerAttribution}{The attribution text of the map tiles.
This is a link to OpenStreetMap with the license CC-BY-SA 2.0 by default.}

\item{options}{A list of map options. See
\url{http://leafletjs.com/reference.html#map-options} for a full list of
options.}
}
\value{
An HTML tag list.
}
\description{
This function is called from \file{ui.R} (or from
\code{\link[shiny]{renderUI}()}); it creates a \code{<div>} that will contain
a Leaflet map.
}

