% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leafletSizingPolicy}
\alias{leafletSizingPolicy}
\title{Leaflet sizing policy}
\usage{
leafletSizingPolicy(
  defaultWidth = "100\%",
  defaultHeight = 400,
  padding = 0,
  browser.fill = TRUE,
  ...
)
}
\arguments{
\item{defaultWidth}{defaults to \code{"100\%"} of the available width}

\item{defaultHeight}{defaults to 400px tall}

\item{padding}{defaults to 0px}

\item{browser.fill}{defaults to \code{TRUE}}

\item{...}{
  Arguments passed on to \code{\link[htmlwidgets:sizingPolicy]{htmlwidgets::sizingPolicy}}
  \describe{
    \item{\code{viewer.defaultWidth}}{The default width used to display the widget
within the RStudio Viewer.}
    \item{\code{viewer.defaultHeight}}{The default height used to display the widget
within the RStudio Viewer.}
    \item{\code{viewer.padding}}{Padding around the widget when displayed in the RStudio
Viewer (defaults to 15 pixels).}
    \item{\code{viewer.fill}}{When displayed in the RStudio Viewer, automatically size
the widget to the viewer dimensions (note that \code{viewer.padding} is
still applied). Default to \code{TRUE}.}
    \item{\code{viewer.suppress}}{Never display the widget within the RStudio Viewer
(useful for widgets that require a large amount of space for rendering).
Defaults to \code{FALSE}.}
    \item{\code{viewer.paneHeight}}{Request that the RStudio Viewer be forced to a
specific height when displaying this widget.}
    \item{\code{browser.defaultWidth}}{The default width used to display the widget
within a standalone web browser.}
    \item{\code{browser.defaultHeight}}{The default height used to display the widget
within a standalone web browser.}
    \item{\code{browser.padding}}{Padding around the widget when displayed in a
standalone browser (defaults to 40 pixels).}
    \item{\code{browser.external}}{When displaying in a browser, always use an external
browser (via \code{\link[=browseURL]{browseURL()}}). Defaults to `FALSE``, which will
result in the use of an internal browser within RStudio v1.1 and higher.}
    \item{\code{knitr.defaultWidth}}{The default width used to display the widget within
documents generated by knitr (e.g. R Markdown).}
    \item{\code{knitr.defaultHeight}}{The default height used to display the widget
within documents generated by knitr (e.g. R Markdown).}
    \item{\code{knitr.figure}}{Apply the default knitr fig.width and fig.height to the
widget when it's rendered within R Markdown documents. Defaults to
\code{TRUE}.}
    \item{\code{fill}}{Whether or not the widget's container should be treated as a fill
item, meaning that its \code{height} is allowed to grow/shrink to fit a fill
container with an opinionated height (see \code{\link[htmltools:bindFillRole]{htmltools::bindFillRole()}} for
more). Examples of fill containers include \code{bslib::card()} and
\code{bslib::card_body_fill()}.}
  }}
}
\value{
An \code{htmlwidgets::sizingPolicy} object
}
\description{
Sizing policy used withing leaflet htmlwidgets.
}
