% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{makeAwesomeIcon}
\alias{makeAwesomeIcon}
\title{Make Awesome Icon}
\usage{
makeAwesomeIcon(icon = NULL, prefix = NULL, markerColor = NULL,
  iconColor = NULL, spin = NULL, extraClasses = NULL)
}
\arguments{
\item{icon}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}

\item{prefix}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}

\item{markerColor}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}

\item{iconColor}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}

\item{spin}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}

\item{extraClasses}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}
}
\description{
Make Awesome Icon
}

