% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{makeIcon}
\alias{makeIcon}
\title{Define icon sets}
\usage{
makeIcon(iconUrl = NULL, iconRetinaUrl = NULL, iconWidth = NULL,
  iconHeight = NULL, iconAnchorX = NULL, iconAnchorY = NULL,
  shadowUrl = NULL, shadowRetinaUrl = NULL, shadowWidth = NULL,
  shadowHeight = NULL, shadowAnchorX = NULL, shadowAnchorY = NULL,
  popupAnchorX = NULL, popupAnchorY = NULL, className = NULL)
}
\arguments{
\item{iconUrl}{the URL or file path to the icon image}

\item{iconRetinaUrl}{the URL or file path to a retina sized version of the
icon image}

\item{iconWidth}{size of the icon image in pixels}

\item{iconHeight}{size of the icon image in pixels}

\item{iconAnchorX}{the coordinates of the "tip" of the icon
(relative to its top left corner, i.e. the top left corner means
\code{iconAnchorX = 0} and \code{iconAnchorY = 0}), and the icon will be
aligned so that this point is at the marker's geographical location}

\item{iconAnchorY}{the coordinates of the "tip" of the icon
(relative to its top left corner, i.e. the top left corner means
\code{iconAnchorX = 0} and \code{iconAnchorY = 0}), and the icon will be
aligned so that this point is at the marker's geographical location}

\item{shadowUrl}{the URL or file path to the icon shadow image}

\item{shadowRetinaUrl}{the URL or file path to the retina sized version of
the icon shadow image}

\item{shadowWidth}{size of the shadow image in pixels}

\item{shadowHeight}{size of the shadow image in pixels}

\item{shadowAnchorX}{the coordinates of the "tip" of the shadow}

\item{shadowAnchorY}{the coordinates of the "tip" of the shadow}

\item{popupAnchorX}{the coordinates of the point from which
popups will "open", relative to the icon anchor}

\item{popupAnchorY}{the coordinates of the point from which
popups will "open", relative to the icon anchor}

\item{className}{a custom class name to assign to both icon and shadow images}
}
\description{
Define icon sets
}
