% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-movezoom.R
\name{moveZoomControl}
\alias{moveZoomControl}
\title{Change the location of the zoom control}
\usage{
moveZoomControl(map, position = c("bottomright", "bottomleft", "topright",
  "topleft"))
}
\arguments{
\item{map}{a map widget object}

\item{position}{the position of the zoom control}
}
\description{
Change the location of the zoom control
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  moveZoomControl()

}
