% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-providers.R
\docType{data}
\name{providers.details}
\alias{providers.details}
\title{Providers Details}
\format{A list of lists (JSON)}
\source{
\url{https://github.com/leaflet-extras/leaflet-providers/blob/master/leaflet-providers.js}
}
\usage{
providers.details
}
\description{
List of all providers with their variations and additional info
}
\keyword{datasets}
