% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-providers.R
\docType{data}
\name{providers.version_num}
\alias{providers.version_num}
\title{Providers version number}
\format{character}
\source{
\url{https://github.com/leaflet-extras/leaflet-providers/blob/master/leaflet-providers.js}
}
\usage{
providers.version_num
}
\description{
Version number of the providers data.
}
\keyword{datasets}
