% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{safeLabel}
\alias{safeLabel}
\title{Sanitize textual labels}
\usage{
safeLabel(label, data)
}
\arguments{
\item{label}{A vector or list of plain characters or HTML (marked by
\code{\link[htmltools]{HTML}}), or a formula that resolves to such a value.}

\item{data}{A data frame over which the formua is evaluated.}
}
\description{
This is a helper function used internally to HTML-escape user-provided
labels. It prevents strings from unintentionally being treated as HTML when
they are intended to be plaintext.
}
\keyword{internal}

