% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setStyle}
\alias{setStyle}
\title{Efficiently style a group that has already been added to the map}
\usage{
setStyle(map, group, styles, label = NULL, offset = 0)
}
\description{
Call with a group and a vector of style lists of length N. The first N
features of the group will be restyled.
}
\examples{
\donttest{
renderLeaflet("map", {
  leaflet() \%>\% addPolygons(data = zones, group = "zones", color = "red")
})
colour = "blue"
styles = lapply(pal(values), function(colour) {list(fillColor=colour, color=colour)})
leafletProxy("map") \%>\%
  setStyle("zones", styles)
}
}
