% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPane.R
\name{updateMapPane}
\alias{updateMapPane}
\title{Update pane on a leaflet map}
\usage{
updateMapPane(map, name, zIndex = NULL, visible = NULL)
}
\arguments{
\item{map}{A \code{leaflet} or \code{mapview} object.}

\item{name}{The name of the new pane (refer to this in \code{leafletOptions}.}

\item{zIndex}{The zIndex of the pane. Panes with higher index are rendered
above panes with lower indices.}

\item{visible}{Should the contents of the pane be visible?
Defaults to \code{TRUE}}
}
\description{
update the parameters for a map pane.
}
\examples{
\donttest{
rand_lng <- function(n = 10) rnorm(n, -93.65, .01)
rand_lat <- function(n = 10) rnorm(n, 42.0285, .01)

random_data <- data.frame(
  lng = rand_lng(50),
  lat = rand_lat(50),
  radius = runif(50, 50, 150),
  circleId = paste0("circle #", 1:50),
  lineId = paste0("circle #", 1:50)
)

leaflet() \%>\%
  addTiles() \%>\%
  # move the center to Snedecor Hall
  setView(-93.65, 42.0285, zoom = 14) \%>\%
  addMapPane("ames_lines", zIndex = 410) \%>\% # shown below ames_circles
  addMapPane("ames_circles", zIndex = 420) \%>\% # shown above ames_lines
  # points above polygons
  addCircles(
    data = random_data, ~lng, ~lat, radius = ~radius, popup = ~circleId,
    options = pathOptions(pane = "ames_circles")
  ) \%>\%
  # lines in 'ames_lines' pane
  addPolylines(
    data = random_data, ~lng, ~lat, color = "#F00", weight = 20,
    options = pathOptions(pane = "ames_lines")
  ) \%>\%
  updateMapPane("ames_circles", visible = FALSE)
}
}
