% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateCoords}
\alias{validateCoords}
\title{Utility function to check if a coordinates is valid}
\usage{
validateCoords(lng, lat, funcName, warn = T)
}
\arguments{
\item{lng}{vector with longitude values}

\item{lat}{vector with latitude values}

\item{funcName}{Name of calling function}

\item{warn}{A boolean. Whether to generate a warning message if there are rows with missing/invalid data}
}
\description{
Utility function to check if a coordinates is valid
}

