% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plugin-heatmap.R
\name{addHeatmap}
\alias{addHeatmap}
\title{Add a heatmap to the map.}
\usage{
addHeatmap(map, latlngs, minOpacity = 0.05, maxZoom = NULL, max = 1,
  radius = 25, blur = 15, gradient = NULL, layerId = NULL)
}
\arguments{
\item{map}{leaflet map to which you would like to add the heatmap.}

\item{latlngs}{matrix of data with latitude in the first column
and longitude in the second.  An optional third column
can provide altitude or intensity.}

\item{minOpacity}{minimum opacity at which the heat will start.}

\item{maxZoom}{zoom level where the points reach maximum intensity
(as intensity scales with zoom).}

\item{max}{maximum point intensity. The default is \code{1.0}.}

\item{radius}{radius of each "point" of the heatmap.  The default is
\code{25}.}

\item{blur}{amount of blur to apply.  The default is \code{15}.
\code{blur=1} means no blur.}

\item{gradient}{manual color gradient.  An example is
\code{gradient = list( '0.4' = 'blue', '0.65' = 'lime', '1'= 'red')}.}

\item{layerId}{optional string identifying this layer.  \code{layerId} can
         be helpful in a dynamic/Shiny situation where you might want to
         remove at some point.}
}
\value{
modified map
}
\description{
Add a heatmap to the map.
}
\examples{
library(leaflet)

# match the heatmap demo from Leaflet.heat
# https://github.com/Leaflet/Leaflet.heat/blob/gh-pages/demo/index.html

# get the data to match exactly
addressPoints <- readLines(
  "http://leaflet.github.io/Leaflet.markercluster/example/realworld.10000.js"
)
addressPoints <- apply(
  jsonlite::fromJSON(
    sprintf("[\%s]",
      paste0(
        addressPoints[4:(length(addressPoints)-1)]
        ,collapse=""
      )
    )
  )
  ,MARGIN = 2
  ,as.numeric
)

# create our heatmap
leaf <- leaflet() \%>\%
  setView( 175.475,-37.87, 12 ) \%>\%
  addHeatmap(addressPoints)

leaf

# customize our heatmap with options
library(RColorBrewer)
pal <- brewer.pal(9,"BuPu")

leaf <- leaflet() \%>\%
  setView( 175.475,-37.87, 12 ) \%>\%
  addHeatmap(
    addressPoints
    ,blur = 50
    # using example
    # https://esri.github.io/esri-leaflet/examples/styling-heatmaps.html
    ,gradient = list(
      '0.2' = '#ffffb2',
      '0.4' = '#fd8d3c',
      '0.6' = '#fd8d3c',
      '0.8' = '#f03b20',
      '1' = '#bd0026'
    )
  )

leaf

# replicate the example provided by
#   http://www.d3noob.org/2014/02/generate-heatmap-with-leafletheat-and.html

earthquakes <- readLines(
  "http://bl.ocks.org/d3noob/raw/8973028/2013-earthquake.js"
)
earthquakes <- apply(
  jsonlite::fromJSON(
    sprintf("[\%s]",
            paste0(
              earthquakes[5:(length(earthquakes)-1)]
              ,collapse=""
            )
    )
  )
  ,MARGIN = 2
  ,as.numeric
)

leaflet() \%>\%
  addTiles() \%>\%
  setView( 174.146, -41.5546, 10 ) \%>\%
  addHeatmap(
    earthquakes,
    radius = 20,
    blur = 15,
    maxZoom = 17
  )

#  using data(quakes)
data(quakes)
quakes_mat <- matrix(t(quakes[,c(1:2,4)]),ncol=3,byrow=T)
leaflet() \%>\%
  addTiles( ) \%>\%
  setView( 178, -20, 5 ) \%>\%
  addHeatmap( quakes_mat, blur = 30, max = 0.05, radius = 30 )

# to remove the heatmap
leaf \%>\% clearHeatmap()
}

