% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-magnifyingGlass.R
\name{addMagnifyingGlass}
\alias{addMagnifyingGlass}
\title{Add a Magnifyer on a Map}
\usage{
addMagnifyingGlass(map, radius = 100, zoomOffset = 3, fixedZoom = -1,
  fixedPosition = FALSE, latLng = c(0, 0), layers = NULL,
  showControlButton = FALSE, layerId = NULL, group = NULL)
}
\arguments{
\item{map}{a map widget object}

\item{radius,zoomOffset,fixedZoom,fixedPosition,latLng,layers;}{see
\url{https://raw.githubusercontent.com/bbecquet/Leaflet.MagnifyingGlass}}

\item{showControlButton}{whether to show a control button or not.}

\item{layerId}{the layer id}

\item{group}{the name of the group this layer belongs to.}
}
\description{
Add a Magnifyer on a Map
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addMagnifyingGlass()
}

