% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-terminator.R
\name{addTerminator}
\alias{addTerminator}
\title{Add a daylight layer on top of the map
see \url{https://github.com/joergdietrich/Leaflet.Terminator}}
\usage{
addTerminator(map, resolution = 2, time = NULL, layerId = NULL,
  group = NULL)
}
\arguments{
\item{map}{a map widget object}

\item{resolution}{the step size at which the terminator points are computed.
The step size is 1°/resolution, i.e. higher resolution values have smaller step sizes and more points in the polygon.
The default value is 2.}

\item{time}{Time}

\item{layerId}{the layer id}

\item{group}{the name of the group this layer belongs to.}
}
\description{
Add a daylight layer on top of the map
see \url{https://github.com/joergdietrich/Leaflet.Terminator}
}
\examples{
library(leaflet)

leaf <- leaflet() \%>\%
  addTiles() \%>\%
  addTerminator()
}

