% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{awesomeIcons}
\alias{awesomeIcons}
\title{Create a list of awesome icon data}
\usage{
awesomeIcons(icon = NULL, prefix = NULL, markerColor = NULL,
  iconColor = NULL, spin = NULL, extraClasses = NULL)
}
\arguments{
\item{icon,prefix,markerColor,iconColor,spin,extraClasses}{see \url{https://github.com/lvoogdt/Leaflet.awesome-markers}}
}
\value{
A list of awesome-icon data that can be passed to the \code{icon} argument of
  \code{\link{addAwesomeMarkers}()}.
}
\description{
An icon can be represented as a list of the form \code{list(icon,
prefix, ...)}. This function is vectorized over its arguments to create a
list of icon data. Shorter argument values will be re-cycled. \code{NULL}
values for these arguments will be ignored.
}

