% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{awesomeIcons}
\alias{awesomeIcons}
\title{Create a list of awesome icon data see
\url{https://github.com/lvoogdt/Leaflet.awesome-markers}}
\usage{
awesomeIcons(icon = "home", library = "glyphicon", markerColor = "blue",
  iconColor = "white", spin = FALSE, extraClasses = NULL)
}
\arguments{
\item{icon}{Name of the icon}

\item{library}{Which icon library. Default \code{'glyphicon'}, other possible
values are \code{'fa'} (fontawesome) or \code{'ion'} (ionicons).}

\item{markerColor}{Possible values are \code{'red'}, \code{'darkred'},
\code{'orange'}, \code{'green'}, \code{'darkgreen'}, \code{'blue'},
\code{'purple'}, \code{'darkpuple'}, \code{'cadetblue'}}

\item{iconColor}{The color to use for the icon itself. Use any CSS-valid
color (hex, rgba, etc.) or a named web color.}

\item{spin}{If \code{TRUE}, make the icon spin (only works when \code{library
= 'fa'})}

\item{extraClasses}{Additional css classes to include on the icon.}
}
\value{
A list of awesome-icon data that can be passed to the \code{icon}
  argument of \code{\link{addAwesomeMarkers}()}.
}
\description{
An icon can be represented as a list of the form \code{list(icon, library,
...)}. This function is vectorized over its arguments to create a list of
icon data. Shorter argument values will be re-cycled. \code{NULL} values for
these arguments will be ignored.
}

