% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filterNULL}
\alias{filterNULL}
\title{remove NULL elements from a list}
\usage{
filterNULL(x)
}
\arguments{
\item{x}{A list whose NULL elements will be filtered}
}
\description{
remove NULL elements from a list
}

