% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gadmCHE}
\alias{gadmCHE}
\title{Administrative borders of Switzerland (level 1)}
\format{
\code{sp::SpatialPolygonsDataFrame}
}
\source{
\url{https://gadm.org}
}
\description{
Administrative borders of Switzerland (level 1)
}
\details{
This dataset comes from \url{https://gadm.org}.
It was downloaded using \code{\link{getData}}.
}
