% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{iconList}
\alias{iconList}
\title{Make icon set}
\usage{
iconList(...)
}
\arguments{
\item{...}{icons created from \code{\link{makeIcon}()}}
}
\description{
Make icon set
}
\examples{

iconSet <- iconList(
  red = makeIcon("leaf-red.png", iconWidth = 32, iconHeight = 32),
  green = makeIcon("leaf-green.png", iconWidth = 32, iconHeight = 32)
)

iconSet[c("red", "green", "red")]
}
