% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{leafletSizingPolicy}
\alias{leafletSizingPolicy}
\title{Leaflet sizing policy}
\usage{
leafletSizingPolicy(defaultWidth = "100\%", defaultHeight = 400,
  padding = 0, browser.fill = TRUE, ...)
}
\arguments{
\item{defaultWidth}{defaults to \code{"100\%"} of the available width}

\item{defaultHeight}{defaults to 400px tall}

\item{padding}{defaults to 0px}

\item{browser.fill}{defaults to \code{TRUE}}

\item{...}{all other arguments supplied to \code{htmlwidgets::\link[htmlwidgets]{sizingPolicy}}}
}
\value{
An \code{htmlwidgets::sizingPolicy} object
}
\description{
Sizing policy used withing leaflet htmlwidgets.  All arguments are passed directly to \code{htmlwidgets::\link[htmlwidgets]{sizingPolicy}}
}
