% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{makeAwesomeIcon}
\alias{makeAwesomeIcon}
\title{Make Awesome Icon}
\usage{
makeAwesomeIcon(icon = NULL, prefix = NULL, markerColor = NULL,
  iconColor = NULL, spin = NULL, extraClasses = NULL)
}
\arguments{
\item{icon}{Name of the icon. Default 'home'}

\item{prefix}{Which icon library. Default 'glyphicon' other values 'fa' (fontawesome) or 'ion' (ionicons)}

\item{markerColor}{Default 'blue', other values ''red', 'darkred', 'orange', 'green', 'darkgreen', 'blue', 'purple', 'darkpuple', 'cadetblue'}

\item{iconColor}{Default 'white'. 'white', 'black' or css code (hex, rgba etc)}

\item{spin}{Make the icon spin. Default 'false' (Font-awesome required)}

\item{extraClasses}{Additional css classes. Default ''.}
}
\description{
Make Awesome Icon
}

