% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-awesomeMarkers.R
\name{makeAwesomeIcon}
\alias{makeAwesomeIcon}
\title{Make Awesome Icon}
\usage{
makeAwesomeIcon(icon = "home", library = "glyphicon",
  markerColor = "blue", iconColor = "white", spin = FALSE,
  extraClasses = NULL, squareMarker = FALSE, iconRotate = 0,
  fontFamily = "monospace", text = NULL)
}
\arguments{
\item{icon}{Name of the icon}

\item{library}{Which icon library. Default \code{"glyphicon"}, other possible
values are \code{"fa"} (fontawesome) or \code{"ion"} (ionicons).}

\item{markerColor}{Possible values are \code{"red"}, \code{"darkred"}, \code{"lightred"}, \code{"orange"},
\code{"beige"}, \code{"green"}, \code{"darkgreen"}, \code{"lightgreen"}, \code{"blue"},
\code{"darkblue"}, \code{"lightblue"}, \code{"purple"}, \code{"darkpurple"}, \code{"pink"},
\code{"cadetblue"}, \code{"white"}, \code{"gray"}, \code{"lightgray"}, \code{"black"}}

\item{iconColor}{The color to use for the icon itself. Use any CSS-valid
color (hex, rgba, etc.) or a named web color.}

\item{spin}{If \code{TRUE}, make the icon spin (only works when \code{library
= "fa"})}

\item{extraClasses}{Additional css classes to include on the icon.}

\item{squareMarker}{Whether to use a square marker.}

\item{iconRotate}{Rotate the icon by a given angle.}

\item{fontFamily}{Used when \code{text} option is specified.}

\item{text}{Use this text string instead of an icon.
argument of \code{\link{addAwesomeMarkers}()}.}
}
\description{
Make Awesome Icon
}
