% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugin-providers.R
\name{providers}
\alias{providers}
\alias{providers.details}
\title{Providers}
\format{A list of characters}
\source{
\url{https://github.com/leaflet-extras/leaflet-providers/blob/master/leaflet-providers.js}
}
\description{
List of all providers with their variations
}
